controller 리펙터링
