+++ server-service/src/main/java/com/playdata/homelesscode/controller/ServerController.java
@@ -83,154 +83,7 @@ public class ServerController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-
-    // 채널 생성
-    @PostMapping("/channels")
-    public ResponseEntity<?> createChannel(ChannelCreateDto dto) {
-
-        Channel result = serverService.createChannel(dto);
-
-        CommonResDto resDto = new CommonResDto(HttpStatus.OK, "채널 생성 성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-    }
-
-    // 채널 목록 조회
-    @GetMapping("/channels")
-    public ResponseEntity<?> getChannel(@RequestParam String id) {
-
-        List<ChannelResponseDto> result = serverService.getChannel(id);
-
-        CommonResDto resDto = new CommonResDto(HttpStatus.OK, "조회 성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-    }
-
-    // 채널 삭제
-    @DeleteMapping("/channels")
-    public ResponseEntity<?> deleteChannel(@RequestParam String id) {
-
-        serverService.deleteChannel(id);
-
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    // 채널 수정
-    @PutMapping("/channels")
-    public ResponseEntity<?> updateChannel(ChannelUpdateDto dto) {
-        Channel result = serverService.updateChannel(dto);
-
-        CommonResDto resDto = new CommonResDto(HttpStatus.OK, "수정성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-
-    }
-
-    // 게시판 생성
-    @PostMapping("/boardList")
-    public ResponseEntity<?> createBoardList(BoardListCreateDto dto) {
-        BoardList result = serverService.createBoardList(dto);
-
-        CommonResDto resDto = new CommonResDto(HttpStatus.OK, "생성 성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-
-    }
-
-    // 게시판 조회
-    @GetMapping("/boardList")
-    public ResponseEntity<?> getBoardList(@RequestParam String id) {
-
-        List<BoardList> result = serverService.getBoardList(id);
-
-        CommonResDto resDto = new CommonResDto(HttpStatus.OK, "조회 성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-    }
-
-    // 게시판 수정
-    @PutMapping("/boardList")
-    public ResponseEntity<?> updateBoardList(BoardListUpdateDto dto) {
-
-        BoardList board = serverService.updateBoardList(dto);
-
-        CommonResDto resDto = new CommonResDto(HttpStatus.OK, "수정 성공", board);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-    }
-
-    // 게시판 삭제
-    @DeleteMapping("/boardList")
-    public ResponseEntity<?> deleteBoardList(@RequestParam String id) {
-
-        log.info("삭제 컨트롤러에여 {} ,", id);
-
-        serverService.deleteBoardList(id);
-
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    // 게시글 생성
-    @PostMapping("/boards")
-    public ResponseEntity<?> createBoards(BoardCreateDto dto) {
-        try {
-            serverService.createBoard(dto);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (IllegalStateException e) {
-            // 이미 삭제된 게시판인 경우
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.CONFLICT);
-        } catch (Exception e) {
-            // 기타 에러 처리
-            return new ResponseEntity<>("게시판 생성 중 문제가 발생했습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    // 게시글 조회
-    @GetMapping("/boards")
-    public ResponseEntity<?> getBoards(BoardSearchDto dto, Pageable pageable) {
-
-        Page<Board> result = serverService.getBoard(dto,pageable);
-
-        CommonResDto resDto = new CommonResDto<>(HttpStatus.OK, "조회 성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-    }
-
-    @DeleteMapping("/boards")
-    public ResponseEntity<?> deleteBoards(BoardDeleteDto dto) {
-
-        serverService.deleteBoard(dto);
-
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-
-    @PutMapping("/boards")
-    public ResponseEntity<?> updateBoards(BoardUpdateDto dto) {
-
-        serverService.updateBoard(dto);
-
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-
-
-    /*
-
-       // 게시글 조회
-    @GetMapping("/boards")
-    public ResponseEntity<?> getBoards(BoardSearchDto dto, Pageable pageable) {
-
-        Page<Board> result = serverService.getBoard(dto,pageable);
-
-        CommonResDto resDto = new CommonResDto<>(HttpStatus.OK, "조회 성공", result);
-
-        return new ResponseEntity<>(resDto, HttpStatus.OK);
-    }
-
-
-    * */
-
+    // 서버 참여 유저 조회
     @GetMapping("/userList")
     public CommonResDto<List<UserReponseInRoleDto>> getUserList(@RequestParam String id ) {
 
@@ -254,6 +107,7 @@ public class ServerController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    // 유저 추방
     @DeleteMapping("/resign")
     public ResponseEntity<?> resignUser(@RequestBody ResignUserDto dto){
         serverService.resignUser(dto);
